RNG (#448)
* lpc55-rng: Update interface to fill an arbitrary buffer with entropy.

This commit changes the lpc55-rng syscall to fill an arbitrary buffer
provided by the caller. This new interface is defined using the IDL.
To simplify access to the RNG we add the rng-api library crate that:
allows consumers to use the familiar 'getrand' and 'rand::os::OsRng'
interfaces to get random bytes from the rng_server task.

* lpc55-rng: Condition RNG output with reseeding ChaCha20 rng.

Chacha20 rng will be reseeded after a threashold number of bytes. This
is currently set to 1 MiB. This is a similar approach to the one taken
in rand::rngs::adapter::ReseedingRng but without the std dependency that
comes from dealing with process forking.

* chips: Uncomment offset & size of rng peripheral / registers on stm32h7.

The RNG has been tested on an h743 & h753.

* stm32h7-startup: Configure RNG to use PLL1 clock for h743 & h753.

* stm32xx-sys-api: Add h743 to features gating the Rng Peripheral.

The RNG hardware is functional on my nucleo h743. This commit is
necessary to enable the Rng clock through the sys api.

* drv: Add RNG driver for stm32h7.

This driver is compatible with h743 and h753 stm32 variants. It uses the
hardware interface described in the ST RM0433 Reference manual, revision
7, section 34. Entropy is taken directly from the hardware RNG and
returned to the caller.

* hiffy: Add RNG interface.

This interface allows humility to request a variable number of bytes
from the RNG. The number of bytes requested by the caller is passed as a
usize on the stack. Hiffy then uses the drv-rng-api crate to fill the
provided rval with entropy from the RNG.